refactor: separate app creation from listening
